fix(lint): isolate hardening tests and policy definitions
