+++ internal/config/environment_config.go
@@ -1,0 +1,8 @@
+package config
+
+// EnvironmentConfig controls startup environment reporting and network facts.
+type EnvironmentConfig struct {
+	Enabled *bool             `toml:"enabled"`
+	Offline bool              `toml:"offline"`
+	Tools   map[string]string `toml:"tools"`
+}
